TNC-011: A diversity of calculators aimed to fill the AlbumStatisticsDto object

+++ src/main/java/amateur/shaobig/tnc/entity/SongMetadata.java
@@ -25,7 +25,7 @@ public class SongMetadata implements Serializable {
     @MapsId
     private Song song;
     private SongType type;
-    private byte mark;
+    private int mark;
 
     public SongMetadata() {
         this.type = SongType.DEFAULT;
